Adding support for binding using zat server
Adding support to allow binding a specific ip address when using zat server. This would be useful for our case when developing using a vagrant/virtual machine, as we tend to bind our tools to 0.0.0.0 when testing and setup our virtual machine to allow access to the virtual machine through a specific ip/host.