Use a fork of zendesk_apps_support
One of the dependencies this gem relies on has some security alerts
(loofah and nokogiri), we have forked and published this dependency
while we wait for the main gem to be patched upstream.

This has a slight impact on dxw having to maintain updates from the
source repo. We should aim to retire our fork as soon as the changes
have been merged upstream.