Add bundler-audit and bundler-audit-fix
Adds tools for auditing vulnerabilities for ZAT. This is important in
order to make sure that we ship secure code.

As part of this change I needed to bump the major version of Thor, which
is a gem used to handle the command line functionality for ZAT.

When running bundler-audit-fix one can run

`bundle exec bundler-audit-fix update .` [1]

This should address vulnerabilities identified by `bundler-audit`.

[1]: https://github.com/nobuyo/bundler-audit-fix